tools/requirements.txt
# LLM Fraud Analyzer Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Async support
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Logging and utilities
python-dotenv>=0.19.0
pyyaml>=6.0

# Optional: For advanced analysis
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0